# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgentFlow AI Agent Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all workflows
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, draft]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow by ID
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    put:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowBody"
      responses:
        "200":
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Workflow deleted
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List available templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [agent, workflow, starter]
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
  /templates/{templateId}/apply:
    post:
      operationId: applyTemplate
      tags: [templates]
      summary: Apply a template to create agent/workflow
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Template applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyTemplateResponse"
  /knowledge-bases:
    get:
      operationId: listKnowledgeBases
      tags: [knowledgeBases]
      summary: List knowledge bases
      responses:
        "200":
          description: List of knowledge bases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeBase"
    post:
      operationId: createKnowledgeBase
      tags: [knowledgeBases]
      summary: Create a knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseBody"
      responses:
        "201":
          description: Knowledge base created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
  /knowledge-bases/{knowledgeBaseId}:
    get:
      operationId: getKnowledgeBase
      tags: [knowledgeBases]
      summary: Get knowledge base details
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Knowledge base details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    delete:
      operationId: deleteKnowledgeBase
      tags: [knowledgeBases]
      summary: Delete a knowledge base
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Knowledge base deleted
  /knowledge-bases/{knowledgeBaseId}/documents:
    get:
      operationId: listDocuments
      tags: [knowledgeBases]
      summary: List documents in a knowledge base
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: addDocument
      tags: [knowledgeBases]
      summary: Add a document to a knowledge base
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentBody"
      responses:
        "201":
          description: Document added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
components:
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, draft]
        tags:
          type: array
          items:
            type: string
        definition:
          type: object
          description: JSON workflow graph definition with nodes and edges
          properties:
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowNode"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowEdge"
        executionCount:
          type: integer
        lastRunAt:
          type: string
          format: date-time
        avgDuration:
          type: number
        successRate:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - definition
        - createdAt
        - updatedAt
    WorkflowNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [trigger, agent, condition, transform, output, delay, loop, parallel, webhook, code, llm_call, api_call, knowledge_query]
        label:
          type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        data:
          type: object
          description: Node-specific configuration data
        agentId:
          type: integer
      required:
        - id
        - type
        - label
        - position
    WorkflowEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        label:
          type: string
        condition:
          type: string
      required:
        - id
        - source
        - target
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        definition:
          type: object
          properties:
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowNode"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowEdge"
      required:
        - name
        - definition
    UpdateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, draft]
        tags:
          type: array
          items:
            type: string
        definition:
          type: object
          properties:
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowNode"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowEdge"
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [agent, workflow, starter]
        icon:
          type: string
        tags:
          type: array
          items:
            type: string
        preview:
          type: object
        popularity:
          type: integer
      required:
        - id
        - name
        - description
        - category
    ApplyTemplateResponse:
      type: object
      properties:
        agentId:
          type: integer
        workflowId:
          type: integer
        message:
          type: string
      required:
        - message
    KnowledgeBase:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        documentCount:
          type: integer
        totalChunks:
          type: integer
        embeddingModel:
          type: string
        status:
          type: string
          enum: [ready, indexing, error]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    CreateKnowledgeBaseBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        embeddingModel:
          type: string
      required:
        - name
    Document:
      type: object
      properties:
        id:
          type: integer
        knowledgeBaseId:
          type: integer
        title:
          type: string
        content:
          type: string
        sourceType:
          type: string
          enum: [text, url, file]
        sourceUrl:
          type: string
        chunkCount:
          type: integer
        status:
          type: string
          enum: [processing, indexed, error]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - knowledgeBaseId
        - title
        - status
        - createdAt
    AddDocumentBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        sourceType:
          type: string
          enum: [text, url, file]
        sourceUrl:
          type: string
      required:
        - title
        - content
...