# lib/api-zod/src/generated/types/addDocumentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AddDocumentBodySourceType } from "./addDocumentBodySourceType";

export interface AddDocumentBody {
  title: string;
  content: string;
  sourceType?: AddDocumentBodySourceType;
  sourceUrl?: string;
}