# lib/api-zod/src/generated/types/knowledgeBase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeBaseStatus } from "./knowledgeBaseStatus";

export interface KnowledgeBase {
  id: number;
  name: string;
  description?: string;
  documentCount?: number;
  totalChunks?: number;
  embeddingModel?: string;
  status: KnowledgeBaseStatus;
  createdAt: Date;
  updatedAt: Date;
}