# lib/api-zod/src/generated/types/updateWorkflowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWorkflowBodyDefinition } from "./updateWorkflowBodyDefinition";
import type { UpdateWorkflowBodyStatus } from "./updateWorkflowBodyStatus";

export interface UpdateWorkflowBody {
  name?: string;
  description?: string;
  status?: UpdateWorkflowBodyStatus;
  tags?: string[];
  definition?: UpdateWorkflowBodyDefinition;
}