# lib/api-zod/src/generated/types/workflowNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowNodeData } from "./workflowNodeData";
import type { WorkflowNodePosition } from "./workflowNodePosition";
import type { WorkflowNodeType } from "./workflowNodeType";

export interface WorkflowNode {
  id: string;
  type: WorkflowNodeType;
  label: string;
  position: WorkflowNodePosition;
  /** Node-specific configuration data */
  data?: WorkflowNodeData;
  agentId?: number;
}